move resolver functions back into api/refs
